refactor: extract history and search handlers from sessions.ts
